Make sure BinaryLogger works with MSBuild 15.3 and later.
We now properly preserve and serialize the new things added in 15.3:
 * ProjectEvaluationStartedEventArgs
 * ProjectEvaluationFinishedEventArgs
 * ProjectImportedEventArgs
 * BuildEventContext.EvaluationId.

Since we only statically reference MSBuild 14.0 we dynamically detect whether the new data is present at runtime using Reflection and gracefully degrade if it's not present.

This now diverges from the BinaryLogger that is checked in in the MSBuild codebase because they have the luxury of referencing 15.3 statically so they don't need reflection.