Add TaskRunner.exe.
A console app that can take a .binlog and execute the specified task from the .binlog using the exact parameters specified in the log. Useful to debug task execution.

Uses MSBuildLocator to control which MSBuild assemblies and toolset to load. Installs a CustomAppDomainManager to install MSBuildLocator assembly resolution for each new appdomain created by the task.

Two modes to locate tasks: by name (runs the first task of that name) or by index (index will be provided by the Viewer).

Logs task output to console.