Massive API breaking change to remove conflicts with MSBuild.
Previously we used to keep a fork of the same public types and namespaces present in MSBuild (such as BinaryLogger, BinaryLogReplayEventSource, etc). It was easier to sync with the MSBuild codebase.

However this presents a problem - if someone is referencing both MSBuild and StructuredLogger.dll then they have a conflict - types with the same namespace and name are present in both places, so it fails the build.

I've decided to move public types which are mirrored from MSBuild into a different namespace, and additionally rename BinaryLogReplayEventSource -> BinLogReader.

Also switching the version from 1.2 to 2.0 as this is a major breaking change.