Don't conflate multiple evaluations in a single preprocess view.
There is an issue with stitching together the preprocessed text for projects where multiple evaluations have resulted in different sets of imports. Suppose one evaluation imported a project and the other didn't. Depending on context the preprocess view should be different and accurately display the actual imports.

Before this change the situation was even worse. We displayed imports in a file regardless of the root project. With this change we at least show the file as evaluated under the matching root project.

This is the first step towards this goal where we separate imports by the root project being evaluated. Next we'll need to consider different evaluations of the same project (like different target frameworks for instance).

See related: #269, #273