Fix SplitterPanel initialization bug with Orientation set
When the `Orientation` property of the SplitterPanel was assigned during initialization of the control we'd
call `UpdateRowColumnInfo()` which in turn retrieves the value of `FirstChildRelativeSize`/`SecondChildRelativeSize`.
That caused the default values for the properties to be used instead of the values set in XAML.

We can remove the explicit `UpdateRowColumnInfo()` call in the `Orientation` setter since it will be called when
`FirstChild` or `SecondChild` is set anyway.