Raise SearchComplete before DisplayResults.
This way Elapsed time will be set correctly.
The problem was that ElapsedTime was set in SearchComplete after it was being used in DisplayResults. So the elapsed label was really showing the result from the previous search, not the current search.

The first search always took 00:00:00.

Technically I don't think we even need SearchComplete.