Initialize Message for some Args in version 13
Starting with file format version 13 some args don't store the Message, so we need to manually recover it to copy what actual MSBuild does. In real MSBuild the Message property on these args is overridden to return the custom string on demand, but we can't do that since we reference older MSBuild, so pretend the message is set.