Improve the exception when a property is not found
FindPropertyInfo is used at three different places but the result was only checked in one place (the SetPropertyValue<T> method). Now the result is checked inside the `FindPropertyInfo` method itself to make sure we have a proper exception rather than a `NullReferenceException` when accessing `propertyInfo.PropertyType` on a null `propertyInfo`.

Also, the message of the exception has been improved to give a hint about the probable root cause of the issue.