Fix WriteProjectItems logic
The WriteItems callback may be called more than once, and we only need to write a zero byte once at the end (this is how the original logic works).

We were writing a zero byte after every callback, which was of course breaking the binary format.

When I originally translated the logic to use Reflection I moved the zero byte logic into the wrong place.