Make Item inherit from NamedNode, not TextNode
This saves on the Text field, which was mostly unused.

We only had a couple of fringe scenarios where both Name and Text were used simultaneously (in Top 10 tasks and C# analyzer report). In both scenarios it is sufficient to concatenate the text into a single field.

This saves us 8 bytes per instance, huge savings for large binlogs with tens of millions of items.