Implement start<, start>, end< and end> search filter syntax
You can use any combination of start<, start>, end< and end>:

start<"2023-11-17 16:56:20.6915121"
start>"2023-11-17 16:56:20.6915121"
end<"2023-11-17 16:56:20.6915121"
end>"2023-11-17 16:56:20.6915121"

The date needs to be in the same format as what Show time and duration right-click context menu shows you for projects, targets and tasks, surrounded with double-quotes. There may be a single space before or after <, but it's not required.

Fixes #722