Disable tasks having searchable subtypes
This creates more problems than it solves. For example searching for "$task msb" would match on "$msbuildtask" and return a result of the shape "$task msbuild msbuildtask". There's no practical benefit to including derived type names as they were invented by this repo and are not widely understood.

Since it's more confusing and produces wrong search results, it's easier to just turn it off. Now instead of searching for "$csctask" you have to search for "$task csc" (or there's two dedicated shortcuts: $csc and $rar).