Show file path in tooltips
When opening source, you might see a tab with a title like `Sdk.targets`. It can be hard to know which SDK this comes from. The full path is shown in the toolbar, but this can be truncated.

This change adds a tooltip to the tab and the toolbar item, showing the full path.