Add "Copy visible subtree" command
The current "Copy subtree" command copies the *entire* subtree, including nodes that aren't visible in the tree given the current expansion state.

This change adds a new command that only copies what's visible in the tree.

My use case was having a long list of targets within a project, and two binlogs (before and after). It's very hard to spot the differences here, so I want to export the target names and compare then in another tool. This command enables that.