Add "Direct references only" option to graph control
The size of the target graph is often so large that related nodes aren't visible within the viewport.

This change adds a checkbox to the toolbar that causes nodes that aren't directly referenced from the selected node. Clicking on one of those nodes will refocus the graph around that node, adding any newly needed nodes and hiding any no longer directly referenced.